error i was getting locally
This is the code in my local editor, left screen is 'main.py' and the right one is the 'database.py'. When I try to run the main.py in terminal it shows me an AttributeError, (the highlighted line) so I guess there might be a possible issue with some of the files I've imported, possibly 'contxtmgr.py' from the traceback that I have got.